requirements.txt
absl-py==0.10.0
asn1crypto==0.24.0
astor==0.8.1
beautifulsoup4==4.6.0
bleach==2.1.2
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.1.2
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.2.3.post1
feedparser==5.2.1
gast==0.4.0
google-pasta==0.2.0
gpg==1.10.0
greenlet==0.4.12
grpcio==1.32.0
h5py==2.10.0
html2text==2018.1.9
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==1.7.0
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.10
joblib==0.16.0
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.2.1
Markdown==3.2.2
MarkupSafe==1.0
meld==3.18.0
mistune==0.8.3
msgpack==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
neovim==0.2.0
netifaces==0.10.4
notebook==5.2.2
notify2==0.3
numpy==1.19.2
olefile==0.45.1
pandas==1.1.2
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.13.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-magic==0.4.16
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
requests==2.18.4
requests-unixsocket==0.1.5
rss2email==3.9
scikit-learn==0.23.2
scipy==1.5.2
scour==0.36
SecretStorage==2.3.1
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
ssh-import-id==5.7
systemd-python==234
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.7
testpath==0.3.1
threadpoolctl==2.1.0
tornado==4.5.3
traitlets==4.3.2
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
